+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Released on Sunday, July 26 2026
 - Added API enhancements to expose asynchronous HTML token source (#1265) @dv00d00
 - Added bounded stream parsing mode (#1263) @dv00d00
 - Added canonicalization of decimal, octal, or hexadecimal IPv4 literals in `Url`
+- Added parsing of / checks for IPv6 literals in `Url`
 - Added support for the CSS column combinator (`||`) selector syntax
 - Added correct identification of `CurrentSource` in `CssMediaElement`